+++ javascript/CholongLim/miniProject/src/main/java/com/example/miniProject/service/jpa/JPABoardService.java
@@ -9,9 +9,9 @@ public interface JPABoardService {
 
     public List<Board> list() throws Exception;
 
-    public Board read(Long boardNo) throws Exception;
+    public Board read(Integer boardNo) throws Exception;
 
-    public void remove(Long boardNo) throws Exception;
+    public void remove(Integer boardNo) throws Exception;
 
     public void modify(Board board) throws Exception;
